Move data out of pages
